added global difference treshold/amount test cases; fixed pattern for pixel amount and also it's representation as Long

+++ src/main/java/org/jboss/lupic/suite/Perception.java
@@ -3,7 +3,7 @@ package org.jboss.lupic.suite;
 public class Perception {
 	short onePixelTreshold = 0;
 	short globalDifferenceTreshold = 0;
-	Short globalDifferencePixelAmount = 0;
+	Long globalDifferencePixelAmount = 0l;
 	Short globalDifferencePercentage = null;
 
 	public short getOnePixelTreshold() {
@@ -22,12 +22,11 @@ public class Perception {
 		this.globalDifferenceTreshold = globalDifferenceTreshold;
 	}
 
-	public Short getGlobalDifferencePixelAmount() {
+	public Long getGlobalDifferencePixelAmount() {
 		return globalDifferencePixelAmount;
 	}
 
-	public void setGlobalDifferencePixelAmount(
-			short globalDifferencePixelAmount) {
+	public void setGlobalDifferencePixelAmount(long globalDifferencePixelAmount) {
 		this.globalDifferencePercentage = null;
 		this.globalDifferencePixelAmount = globalDifferencePixelAmount;
 	}
